move Platform from public api to runtime
